requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
futures==3.0.5
goslate==1.5.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysql-connector==2.2.9
numpy==1.26.4
openai==1.35.10
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
typing==3.7.4.3
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
